+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/afericao/relatorios/AfericaoRelatorioDao.java
@@ -1,6 +1,7 @@
 package br.com.zalf.prolog.webservice.frota.pneu.afericao.relatorios;
 
 import br.com.zalf.prolog.webservice.commons.report.Report;
+import br.com.zalf.prolog.webservice.frota.pneu.afericao.relatorios._model.AfericaoExportacaoProtheus;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.OutputStream;
@@ -13,7 +14,6 @@ import java.util.List;
  * @author Thais Francisco (https://github.com/thaisksf)
  */
 public interface AfericaoRelatorioDao {
-
     /**
      * Método para buscar o relatório de cronograma das aferições de placas em CSV.
      *
@@ -63,4 +63,22 @@ public interface AfericaoRelatorioDao {
     Report getDadosGeraisAfericoesReport(@NotNull final List<Long> codUnidades,
                                          @NotNull final LocalDate dataInicial,
                                          @NotNull final LocalDate dataFinal) throws Throwable;
+
+    /**
+     * Método utilizado para buscar um objeto no padrão do Protheus para possibilitar a importação das aferições dentro
+     * do sistema Protheus. Nesta busca, permitimos o filtro de data. Contudo, se qualquer uma das datas forem nulas,
+     * o filtro de data será desconsiderado e buscaremos as aferições para todas as datas.
+     *
+     * @param codUnidades uma lista de unidades a serem usadas no filtro da busca.
+     * @param codVeiculos uma lista de veículos a serem usados no filtro da busca.
+     * @param dataInicial a data inicial a ser usada no filtro, podendo ser nula.
+     * @param dataFinal   a data final a ser usada no filtro, podendo ser nula.
+     * @return um objeto no padrão do Protheus com as informações das aferições.
+     * @throws Throwable se qualquer erro ocorrer.
+     */
+    @NotNull
+    List<AfericaoExportacaoProtheus> getExportacaoAfericoesProtheus(@NotNull final List<Long> codUnidades,
+                                                                    @NotNull final List<Long> codVeiculos,
+                                                                    @NotNull final LocalDate dataInicial,
+                                                                    @NotNull final LocalDate dataFinal) throws Throwable;
 }